+++ Source/POETradeHelper.ItemSearch.Contract/Models/ItemStats/MinMaxValueItemStat.cs
@@ -15,8 +15,8 @@
             }
         }
 
-        public int MinValue { get; set; }
+        public decimal MinValue { get; set; }
 
-        public int MaxValue { get; set; }
+        public decimal MaxValue { get; set; }
     }
 }
